Add CLANG_VERSION_SUFFIX to extend the version string
LLVM uses SVN for version control so I would like some way to show the
revision at the time of build in the version string without needing to
use SVN locally (because using git is just much more natural and easy
with my script). I've been setting the patch level to it but that
confuses some people and potentially programs relying on patch level.

Add CLANG_VERSION_SUFFIX like LLVM_VERSION_SUFFIX and append it to
the '--version' display.

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>